# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/guilds":{"get":{"tags":["guilds"],"summary":"Get","description":"Retorna uma lista de guildas disponíveis.\n\nReturns\n-------\nlist[IGuild]\n    As guildas.","operationId":"get_api_guilds_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/IGuild"},"type":"array","title":"Response Get Api Guilds Get"}}}},"404":{"description":"Not found"}}}},"/api/guilds/":{"get":{"tags":["guilds"],"summary":"Get Guild","description":"Retorna uma guilda a partir de um ID.\n\nParameters\n----------\nrequest : Request\n    _description_\nidx : str\n    ID da Guilda.\n\nReturns\n-------\nIGuild\n    _description_","operationId":"get_guild_api_guilds__get","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/IGuild"},{"type":"null"}],"title":"Response Get Guild Api Guilds  Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/music/list":{"get":{"tags":["guilds"],"summary":"Get Music List","description":"Retorna a lista de músicas de uma Guilda (Incluindo a música atual).\n\nParameters\n----------\nrequest : Request\n    _description_\nidx : str\n    Guild ID.\n\nReturns\n-------\nlist[IMusic]\n    _description_","operationId":"get_music_list_api_guilds_music_list_get","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/IMusic"},"title":"Response Get Music List Api Guilds Music List Get"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/state":{"get":{"tags":["guilds"],"summary":"Get Music State","description":"Get the current music state for a guild.\n\nReturns\n-------\nIMusicState\n    Complete music state.","operationId":"get_music_state_api_guilds_state_get","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IMusicState"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/queue":{"post":{"tags":["guilds"],"summary":"Add To Queue","description":"Add a song to the queue.\n\nRaises\n------\nHTTPException\n    If the URL is empty.","operationId":"add_to_queue_api_guilds_queue_post","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}},{"name":"url","in":"query","required":true,"schema":{"type":"string","title":"Url"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"null","title":"Response Add To Queue Api Guilds Queue Post"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/canvas":{"get":{"tags":["guilds"],"summary":"Get Canvas","description":"Obtém os dados do canvas para uma guilda específica.\n\nParameters\n----------\nrequest : Request\n    Requisição FastAPI\nidx : str\n    ID da guilda\n\nReturns\n-------\nCanvasData\n    Dados do canvas da guilda","operationId":"get_canvas_api_guilds_canvas_get","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CanvasData"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["guilds"],"summary":"Update Canvas","description":"Atualiza os dados do canvas para uma guilda específica.\n\nParameters\n----------\nrequest : Request\n    Requisição FastAPI\nidx : str\n    ID da guilda\ncanvas_data : CanvasData\n    Novos dados do canvas\n\nReturns\n-------\ndict[str, str]\n    Mensagem de confirmação","operationId":"update_canvas_api_guilds_canvas_post","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CanvasData"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"},"title":"Response Update Canvas Api Guilds Canvas Post"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/canvas/images":{"post":{"tags":["guilds"],"summary":"Upload Image","description":"Faz upload de uma imagem para o canvas.\n\nArgs:\n    image_data (ImageData): Dados da imagem em base64.\n    \nReturns:\n    ImageReference: Referência para a imagem armazenada.","operationId":"upload_image_api_guilds_canvas_images_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageData"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageReference"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/canvas/images/{image_id}":{"get":{"tags":["guilds"],"summary":"Get Image","description":"Obtém uma imagem pelo seu ID.\n\nArgs:\n    image_id (str): ID da imagem.\n    file_type (str): Tipo de arquivo da imagem.\n    \nReturns:\n    Response: Conteúdo binário da imagem.","operationId":"get_image_api_guilds_canvas_images__image_id__get","parameters":[{"name":"image_id","in":"path","required":true,"schema":{"type":"string","title":"Image Id"}},{"name":"file_type","in":"query","required":true,"schema":{"type":"string","title":"File Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/skip":{"post":{"tags":["guilds"],"summary":"Skip Music","description":"Skip the current music.","operationId":"skip_music_api_guilds_skip_post","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"null","title":"Response Skip Music Api Guilds Skip Post"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/pause":{"post":{"tags":["guilds"],"summary":"Pause Music","description":"Pause the current music.","operationId":"pause_music_api_guilds_pause_post","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"null","title":"Response Pause Music Api Guilds Pause Post"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/resume":{"post":{"tags":["guilds"],"summary":"Resume Music","description":"Resume the current music.","operationId":"resume_music_api_guilds_resume_post","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"null","title":"Response Resume Music Api Guilds Resume Post"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/stop":{"post":{"tags":["guilds"],"summary":"Stop Music","description":"Stop the current music.","operationId":"stop_music_api_guilds_stop_post","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"null","title":"Response Stop Music Api Guilds Stop Post"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/loop":{"post":{"tags":["guilds"],"summary":"Loop Music","description":"Alterna o modo de loop.","operationId":"loop_music_api_guilds_loop_post","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ILoopRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"null","title":"Response Loop Music Api Guilds Loop Post"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/guilds/voice/connect":{"post":{"tags":["guilds"],"summary":"Connect To Voice","description":"Connect to a voice channel.","operationId":"connect_to_voice_api_guilds_voice_connect_post","parameters":[{"name":"idx","in":"query","required":true,"schema":{"type":"string","title":"Idx"}},{"name":"channel_id","in":"query","required":true,"schema":{"anyOf":[{"type":"string"},{"type":"integer"}],"title":"Channel Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"null","title":"Response Connect To Voice Api Guilds Voice Connect Post"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/system/status":{"get":{"tags":["system"],"summary":"Get System Status","description":"Obtém o status atual do sistema.\n\nRetorna informações sobre CPU, memória, discos e uptime.\n\nReturns\n-------\nSystemStatus\n    Objeto contendo estatísticas do sistema.","operationId":"get_system_status_api_system_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SystemStatus"}}}},"404":{"description":"Not found"}}}},"/api/system/minecraft":{"get":{"tags":["system"],"summary":"Get Minecraft Status","description":"Obtém o status do servidor Minecraft.\n\nVerifica se o servidor está rodando e obtém informações como\nversão, jogadores online e uso de recursos.\n\nReturns\n-------\nMinecraftStatus\n    Status do servidor Minecraft.","operationId":"get_minecraft_status_api_system_minecraft_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MinecraftStatus"}}}},"404":{"description":"Not found"}}}},"/api/system/top":{"get":{"tags":["system"],"summary":"Get Top Processes","description":"Obtém informações sobre os processos mais ativos do sistema.\n\nSimilar ao comando 'top' do Linux, retorna uma lista dos processos\nque estão consumindo mais recursos.\n\nReturns\n-------\nTopProcesses\n    Lista dos processos mais ativos.","operationId":"get_top_processes_api_system_top_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TopProcesses"}}}},"404":{"description":"Not found"}}}},"/api/system/top/image":{"get":{"tags":["system"],"summary":"Get Top Image","description":"Retorna uma imagem com a saída do comando 'top'.\n\nGera uma imagem contendo a saída formatada do comando 'top',\nsemelhante ao que é feito no comando do Discord.\n\nReturns\n-------\nStreamingResponse\n    Imagem PNG contendo a saída do 'top'.","operationId":"get_top_image_api_system_top_image_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"404":{"description":"Not found"}}}},"/api/status":{"get":{"tags":["api"],"summary":"Bot Status","description":"Check the bot's status via a FastAPI endpoint.\n\nReturns\n-------\nIStatus\n    Status.","operationId":"bot_status_api_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IStatus"}}}}}}}},"components":{"schemas":{"CanvasData":{"properties":{"elements":{"items":{"type":"object"},"type":"array","title":"Elements"},"app_state":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"App State"},"version":{"type":"integer","title":"Version","default":0}},"type":"object","title":"CanvasData","description":"Dados do canvas para uma guilda."},"DiskInfo":{"properties":{"path":{"type":"string","title":"Path"},"total":{"type":"number","title":"Total"},"used":{"type":"number","title":"Used"},"free":{"type":"number","title":"Free"},"percent":{"type":"number","title":"Percent"}},"type":"object","required":["path","total","used","free","percent"],"title":"DiskInfo","description":"Informações de um disco."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IGuild":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"approximate_member_count":{"type":"integer","title":"Approximate Member Count"},"icon":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Icon"},"voice_channels":{"anyOf":[{"items":{"$ref":"#/components/schemas/IVoiceChannel"},"type":"array"},{"type":"null"}],"title":"Voice Channels"},"text_channels":{"anyOf":[{"items":{"$ref":"#/components/schemas/ITextChannel"},"type":"array"},{"type":"null"}],"title":"Text Channels"}},"type":"object","required":["id","name","description","approximate_member_count","icon","voice_channels","text_channels"],"title":"IGuild","description":"Guild model."},"ILoopRequest":{"properties":{"mode":{"$ref":"#/components/schemas/LoopMode"}},"type":"object","required":["mode"],"title":"ILoopRequest","description":"Request to toggle loop"},"IMusic":{"properties":{"title":{"type":"string","title":"Title"},"url":{"type":"string","title":"Url"},"thumbnail":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail"},"duration":{"type":"number","title":"Duration"},"artists":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Artists"},"album":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Album"},"categories":{"items":{"type":"string"},"type":"array","title":"Categories"},"release_year":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Release Year"}},"type":"object","required":["title","url","thumbnail","duration","artists","album","categories","release_year"],"title":"IMusic","description":"Music data model."},"IMusicState":{"properties":{"queue":{"items":{"$ref":"#/components/schemas/IMusic"},"type":"array","title":"Queue"},"loop_mode":{"$ref":"#/components/schemas/LoopMode"},"progress":{"type":"number","title":"Progress"},"current_voice_channel":{"anyOf":[{"$ref":"#/components/schemas/IVoiceChannel"},{"type":"null"}]},"paused":{"type":"boolean","title":"Paused","default":false},"playing":{"type":"boolean","title":"Playing","default":false}},"type":"object","required":["queue","loop_mode","progress","current_voice_channel"],"title":"IMusicState","description":"Model for sending all the Music state of a guild."},"IStatus":{"properties":{"status":{"type":"string","enum":["online","offline"],"title":"Status"}},"type":"object","required":["status"],"title":"IStatus","description":"Estado atual do Bot."},"ITextChannel":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["id","name"],"title":"ITextChannel","description":"Text channel model."},"IVoiceChannel":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"members":{"items":{"type":"string"},"type":"array","title":"Members"}},"type":"object","required":["id","name","members"],"title":"IVoiceChannel","description":"Voice channel model."},"ImageData":{"properties":{"content":{"type":"string","title":"Content"},"file_type":{"type":"string","title":"File Type"}},"type":"object","required":["content","file_type"],"title":"ImageData","description":"Dados de uma imagem para armazenamento."},"ImageReference":{"properties":{"image_id":{"type":"string","title":"Image Id"},"file_type":{"type":"string","title":"File Type"}},"type":"object","required":["image_id","file_type"],"title":"ImageReference","description":"Referência a uma imagem armazenada."},"LoopMode":{"type":"integer","enum":[0,1,2],"title":"LoopMode","description":"Enum que representa o modo de loop."},"MinecraftPlayer":{"properties":{"name":{"type":"string","title":"Name"},"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"}},"type":"object","required":["name"],"title":"MinecraftPlayer","description":"Informações de um jogador de Minecraft."},"MinecraftStatus":{"properties":{"is_running":{"type":"boolean","title":"Is Running","description":"Se o servidor está rodando"},"version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Version"},"online_players":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Online Players"},"max_players":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Players"},"players_list":{"items":{"$ref":"#/components/schemas/MinecraftPlayer"},"type":"array","title":"Players List","default":[]},"cpu_percent":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Cpu Percent"},"memory_mb":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Memory Mb"}},"type":"object","required":["is_running"],"title":"MinecraftStatus","description":"Status do servidor Minecraft."},"ProcessInfo":{"properties":{"pid":{"type":"integer","title":"Pid"},"name":{"type":"string","title":"Name"},"cpu_percent":{"type":"number","title":"Cpu Percent"},"memory_percent":{"type":"number","title":"Memory Percent"},"command":{"type":"string","title":"Command"}},"type":"object","required":["pid","name","cpu_percent","memory_percent","command"],"title":"ProcessInfo","description":"Informações de um processo do sistema."},"SystemStatus":{"properties":{"cpu_percent":{"type":"number","title":"Cpu Percent","description":"Percentual de uso de CPU"},"memory_percent":{"type":"number","title":"Memory Percent","description":"Percentual de uso de memória"},"memory_total":{"type":"number","title":"Memory Total","description":"Total de memória em GB"},"memory_used":{"type":"number","title":"Memory Used","description":"Memória usada em GB"},"uptime":{"type":"string","title":"Uptime","description":"Tempo de atividade do sistema"},"disks":{"items":{"$ref":"#/components/schemas/DiskInfo"},"type":"array","title":"Disks","description":"Informações de discos"}},"type":"object","required":["cpu_percent","memory_percent","memory_total","memory_used","uptime","disks"],"title":"SystemStatus","description":"Status do sistema com informações de recursos."},"TopProcesses":{"properties":{"timestamp":{"type":"string","title":"Timestamp","description":"Timestamp da coleta"},"processes":{"items":{"$ref":"#/components/schemas/ProcessInfo"},"type":"array","title":"Processes","description":"Lista de processos"}},"type":"object","required":["timestamp","processes"],"title":"TopProcesses","description":"Lista dos processos mais ativos do sistema."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}